Move card components to standalone component
